The Essential Role of Marketing Operations in Business Success
This module examined the components of marketing operations and how marketing operations contribute to the failure or success of a business. From analytics, competitor, customer and market intelligence, strategic planning, budgeting, pricing, project management, talent, team building, and integration of technology for analytics, workflow, and automation, I  acquired a comprehensive understanding of the full range and function of roles within Marketing Operations.