Brand Touchpoints Across All Media
In the previous module, I learned a framework for understanding consumer psychology and behavior and how marketers use this insight to develop marketing strategies. My next step is to be able to identify the marketing channels and brand experiences the customer encounters when coming into contact with the brand. These experiences are called brand touchpoints and include contact and interactions in the customer’s journey that impact awareness, discovery, competitive analysis, purchase, and engagement. We will also explore the nuances of media channels and how perceptions are formed by interactions with these channels. Finally, I will explore brand loyalty, which may be developed if touchpoint experiences are well executed.